Allow client assigned IDs to be purely numeric in JPA server if configured to do so

+++ hapi-fhir-jpaserver-base/src/test/java/ca/uhn/fhir/jpa/dao/r4/FhirResourceDaoR4SearchCustomSearchParamTest.java
@@ -146,8 +146,8 @@ public class FhirResourceDaoR4SearchCustomSearchParamTest extends BaseJpaR4Test 
 
 		mySearchParameterDao.create(fooSp, mySrd);
 
-		assertEquals(1, myResourceReindexingSvc.forceReindexingPass().intValue());
-		assertEquals(0, myResourceReindexingSvc.forceReindexingPass().intValue());
+		assertEquals(1, myResourceReindexingSvc.forceReindexingPass());
+		assertEquals(0, myResourceReindexingSvc.forceReindexingPass());
 
 	}
